Reorganized code; support for more private operations

+++ data/synthetic/gen_synthetic_gaussian.py
@@ -22,7 +22,7 @@ def main():
         for v in values:
             f.write('%0.02f' % v + '\n')
 
-    schema = [{'name':colname, 'kind':'Integer'}]
+    schema = [{'name':colname, 'kind':'Double'}]
     with open('gaussian.schema', 'w') as f:
         f.write(json.dumps(schema))
